requirements.txt
python-telegram-bot==13.15
paramiko>=3.4.0
flask==2.3.3
werkzeug==2.3.7
pywinrm==0.4.3
requests==2.31.0
cryptography>=41.0.8
